feat: UI 디자인

+++ clicker/Assets/01.Scripts/Ingame/Click/ClickInfo.cs
@@ -3,6 +3,6 @@
 public class ClickInfo
 {
     public EClickType Type;
-    public int Damage;
+    public double Damage;
     public Vector2 Position;
 }
